refactor(admin): extract OrganizationsTab into its own file
